oop/requirements.txt
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Django==1.10
django-crontab==0.7.1
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
joblib==0.14.0
numpy==1.17.4
pycparser==2.19
pytz==2019.3
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
urllib3==1.25.7
